+++ src/components/ListOfNotes/ListOfNotes.js
@@ -1,25 +1,6 @@
-import React, { useEffect, useState } from 'react';
-import localforage from 'localforage';
+import React from 'react';
 
-function ListOfNotes({ data, listName, onRemove, passIdInMD }) {
-  const passId = (id) => {
-    passIdInMD(id);
-  };
-  const removeItem = (id) => {
-    localforage
-      .removeItem(id)
-      .then(function () {
-        // Run this code once the key has been removed.
-        console.log('Key is cleared!');
-        passId(id);
-      })
-      .catch(function (err) {
-        // This code runs if there were any errors
-        console.log(err);
-      });
-  };
-  // console.log(data);
-
+function ListOfNotes({ data, listName, passId }) {
   return (
     <div
       className="list-of-notes"
@@ -42,7 +23,7 @@ function ListOfNotes({ data, listName, onRemove, passIdInMD }) {
           >
             <div className="note-title">{key}</div>
             <div className="note-btn">
-              <button onClick={() => removeItem(key)}>Delete</button>
+              <button onClick={() => passId(key)}>Delete</button>
             </div>
           </div>
         ))}
