+++ src/components/ListOfNotes/ListOfNotes.md
@@ -1,11 +1,9 @@
 ```jsx
 import React from 'react';
+
 import { ListOfNotes, useData } from '@texttree/notepad-rcl';
 
-const { notes, setNotes, updateNotes } = useData();
-<ListOfNotes
-  data={notes}
-  listName="List of Notes"
-  passIdInMD={(id) => updateNotes(id)}
-/>;
+const { notes, setNotes, removeItem } = useData();
+
+<ListOfNotes data={notes} listName="List of Notes" passId={(id) => removeItem(id)} />;
 ```
